refactor(question): :art: Refactored structure of question service.
