+++ questionservice/Dockerfile
@@ -11,8 +11,8 @@ COPY package*.json ./
 RUN npm install --ignore-scripts --omit=dev
 
 # Copy the app source code to the working directory
-COPY ./question-service.js ./ 
-COPY ./queries.js ./
+COPY ./src/index.js ./ 
+COPY ./src/queries.js ./
 # TODO: Maybe move the src files to a src folder
 
 # Expose the port the app runs on
